rest ok

+++ common/user_info.go
@@ -4,5 +4,5 @@ type UserInfo struct {
 	Id       string `json:"id" description:"identifier of the user"`
 	Password string `json:"password" description:"password of user"`
 	Name     string `json:"name" description:"name of the user"`
-	Created  int    `json:"created" description:"created time`
+	Created  int64    `json:"created" description:"created time`
 }
